+++ src/sample/model/IGeneradorAleatorio.java
@@ -3,7 +3,7 @@ package sample.model;
 import java.util.ArrayList;
 
 public interface IGeneradorAleatorio {
-    public int getSemilla();
+    public long getSemilla();
 
     public float GenerarAleatorio();
